Tests: Dicts are logged via Log.debug which logs dicts with sorted items
Dicts are unordered in versions below Py36. Logging the dicts may
generate a different order of items every time, which restricts the
log comparison with test logs. Hence, LOG.debug is overridden in
to log dicts with sorted items in the tests and all the dicts are
logged via LOG.debug in the module.