Add  user ids argument for snapshot permissions
When an image gets pushed a snapshot will be created if one
doesn't already exists. The current workflow doesn't grant any
"CreateVolumePermission" permissions, though it is expected
certain accounts should have them.

This change introduces an argument to take in a list of user ids,
which will then be passed to the publish function in the cloudimg
library.