Change --snapshot-account-ids parsing [RHELDST-11233]
When I added in the --snapshot-account-ids option initially, I had
thought pubtools-ami was primarily used via commandline. It turns out
that it's actually primarily called via pub and values are injected
there. This change makes the option a list of comma separated ids, which
will be easier to store and pass from pub.