Remove defaults for snapshot-account-ids [RHELDST-11233]
It has been requested that we move the default values for
snapshot-account-ids to the admin portal. This is because we want to
keep this script as generic as possible and remove the need for PRs when
updating default values.

It was also requested to remove duplicate values from
snapshot-account-ids input. When default values are included, they're
ignored by the AWS API. It's more a point of proper input sanitation
and will keep us safe incase the API changes in the future.