Share AMIs publicly based on public_image flag [RHELDST-16452]
pushsource and cloud-image-tools recently implemented the support for
a flag designating whether the image should be released publicly (i.e.
shared to group "all") which is set in staged metadata instead of
the hardcoded rule for hourly images in pubtools-ami. This commit
implements the remaining part which controls the actual behavior
responsible for sharing the images publicly during pub image pushes.

As backwards-compatible behavior, the hourly images without any
public_image flag set in the metadata are still considered to be public
and handled as such.