Fix setting AMI push item name
Previously in case of AWS upload failure,
the name of pushitem stored in images.json was set
to `None` which is invalid because `name` field is not optional and
doesn't allow `None` value.
This behavior broke workflow for new pubtools-ami-delete command
where we weren't able to load push items from Pub, if one of
pushitems had `name` set as None.

With this change `name` field will always be set to the name of the
image even if upload to AWS fails. The value of this field is derived
from provided metadata and doesn't rely on data from AWS.